refactor helper methods
